refactor genre id type to long and add repository tests
- Change `Genre` ID from `Int` to `Long` in domain models, repositories, and `GenresTable`
- Update `GenreRepository.saveGenre` to return the complete `Genre` object instead of its ID
- Refactor `MovieGenreTable` to use a standard Exposed `Table` with direct `insert` statements instead of `CompositeEntity`
- Fix indentation for `id BIGSERIAL PRIMARY KEY` definitions across all SQL migration scripts
- Add integration tests in `MovieRepositoryImplTest` to verify saving genres and movies

+++ src/main/kotlin/features/movies/domain/repository/MovieRepository.kt
@@ -9,5 +9,5 @@ interface MovieRepository {
     suspend fun getMovieById(movieId: Long): DataResult<Movie>
     suspend fun updateMovie(movie: Movie): DataResult<Long>
     suspend fun deleteMovie(id: Long): DataResult<Int>
-    suspend fun getMoviesByGenre(genreId: Int, limit: Int, offset: Long): DataResult<List<Movie>>
+    suspend fun getMoviesByGenre(genreId: Long, limit: Int, offset: Long): DataResult<List<Movie>>
 }
